+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 - Address book: added business fields in detail panel of orginisation
 - Address book: Create personal address books for each user
+- Users: Only show module tabs in user dialog where this user has access to
+- Notes: Default note book setting and every user gets a personal note book
+- Sync: settings use default notebook and address book by default
 
 19-09-2019 6.4.57
 - Custom fields: Query language works with text for select and multiselect